pyproject.toml
[project]
name = "6-coockies-auth"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-multipart (>=0.0.20,<0.0.21)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "attrs (>=25.1.0,<26.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
